refactor: Introduce "tracing" plugin and switch request pipeline to use it.
This commit introduces a `plugin` (named) export from the `apollo-tracing`
package, which uses the new "plugin" hooks rather than the previous
implementation (exported as `TracingExtension`) which uses the
soon-to-be-deprecated "extensions" hooks.  The functionality is intended to
be identical, in spirit.

Since the delta of the commits was otherwise confusing, I've left the
`TracingExtension` present and exported and will remove it in a subsequent
commit.

Briefly summarizing what the necessary changes were:

1. We no longer use a class instance to house the extension, which was
   necessitated by the `graphql-extensions` API.  This means that uses of
   `this` have been replaced with function scoped variables by the same name.
2. The logic which actually does the formatting (previously handled by the
   `format` method in `graphql-extension`, now takes place within the plugin
   API's `willSendResponse` method.